saving_coordinate_into_DB

+++ db/migrate/20190926083918_add_lnt_and_lat_to_foods.rb
@@ -1,6 +1,6 @@
 class AddLntAndLatToFoods < ActiveRecord::Migration[5.2]
   def change
-    add_column :foods, :lat, :decimal
-    add_column :foods, :lng, :decimal
+    add_column :foods, :lat, :float
+    add_column :foods, :lng, :float
   end
 end
